[256936] Support for Offline Mode https://bugs.eclipse.org/bugs/show_bug.cgi?id=256936

+++ plugins/org.eclipse.net4j.db.h2/src/org/eclipse/net4j/db/h2/H2Adapter.java
@@ -61,7 +61,7 @@ public class H2Adapter extends DBAdapter
       return "VARCHAR"; //$NON-NLS-1$
 
     case NUMERIC:
-      return "INT"; //$NON-NLS-1$
+      return "DECIMAL"; //$NON-NLS-1$
 
     case LONGVARBINARY:
     case VARBINARY:
